Create BarcodeScanner
This page triggers and scans the barcode of the product, works with mobile and pc devices. Only issue is that you may have to refresh the page in order to scan a new product.

Only shows ingredients, if you have the data you could easily tweak it to show nutrition per servings, allergens, etc.